[yarp] Fix regression in yarpmanager fixture manager
The regression was introduced by https://github.com/robotology/yarp/commit/4403ebbe380852d764df8eb33dab08eebec4d3a8
that changed the API of the manager-related yarp class, preventing to load any fixture using yarpmanager.

The fix was inspired by the similar pieces of code found in yarpmanager.

Bumping the required YARP version because we need a version of YARP
with the API with the latest change.