Fix project() VERSION issue on whole project
To avoid generating VERSION policy issue CMP0048 while
configuring the project, the project root CMakeLists and the RTF
CMakeLists have been modified.
In particular, in RTF CMakeLists.txt the project(RTF) call has been
removed because it was redundant and, as a conseguence, the cause of the
version policy issue.
As a conseguence, several variable in RTF CMakeLists.txt have been
modified to account for the new path of the RTF project which is
located in src/rtf/* folder and not in the root one where
project(RTF) is originally defined.

Fixes #48